# [envsec] Improve performance