--- a/.github/workflows/monorepo-go.yml
+++ b/.github/workflows/monorepo-go.yml
@@ -19,7 +19,7 @@ jobs:
         uses: actions/checkout@v3
 
       - name: Install devbox
-        uses: jetpack-io/devbox-install-action@v0.5.0
+        uses: jetpack-io/devbox-install-action@v0.6.0
 
       - name: Go modules should be up-to-date
         run: |